Added a project for boundary tests.
AKA Acceptance Tests or Subcutaneous Tests. The purpose of these tests is
to verify that the external boundary of the HTTP API works as intended -
e.g. that it responds at all, that the returned HTTP bodies, as well as
the accepted POST/PUT payloads, have the correct shape (basically that the
JSON looks correct).

Added Xunit.Extensions NuGet package to Boundary Tests.